feat: refactor notification components for improved structure and readability
