+++ client/src/components/layout/NoticationCenter/NotificationMenuHeader.tsx
@@ -1,0 +1,12 @@
+import { Typography } from '@mui/material';
+
+const NotificationMenuHeader = () => (
+  <div style={{ borderBottom: '1px solid #e0e0e0', padding: '8px 16px' }}>
+    <Typography variant="subtitle1" fontWeight={600}>
+      Notifications
+    </Typography>
+  </div>
+);
+
+export default NotificationMenuHeader;
+
